% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isnullptr.R
\name{isnullptr}
\alias{isnullptr}
\title{Check if an `externalptr` is a null pointer}
\usage{
isnullptr(pointer)
}
\arguments{
\item{pointer}{object of type `externalptr`}
}
\value{
TRUE if null pointer, FALSE otherwise
}
\description{
\code{isnullptr} checks if an object of type `externalptr` is a null pointer.
}
\examples{
isnullptr(new("externalptr"))
isnullptr("any_data")
}
